+++ src/theme/overrides/MuiChip.js
@@ -1,0 +1,7 @@
+const MuiChip = {
+  root: {
+    textTransform: 'lowercase',
+  },
+}
+
+export default MuiChip