+++ CHANGELOG.md
@@ -12,6 +12,7 @@ All non-core API/package changes (i.e. changes that do not affect the package de
 - `<ThemeProvider>` - Added custom `ThemeProvider` that uses `react-labs` theme by default.
 - `Theme` - Subtitle and body variations added.
 ### Fixed
+- `<DefaultTheme>` - Fixed theme object export.
 - `<Typography />` - Fixed children prop-type typo.
 - `<Typography />` - secondary color configuration prop added.
 - `Theme` - DefaultTheme not found warning fixed.